Rename packages in runtimeconfig module
add jacoco gradle plugin
