+++ spring-funk-runtimeconfig/src/main/kotlin/com/github/wakingrufus/funk/RuntimeConfig.kt
@@ -1,4 +1,4 @@
-package com.github.wakingrufus.springdsl.runtimeconfig
+package com.github.wakingrufus.funk
 
 import java.util.function.Supplier
