+++ spring-funk-runtimeconfig/src/main/kotlin/com/github/wakingrufus/funk/RuntimeConfigDslInitializer.kt
@@ -1,4 +1,4 @@
-package com.github.wakingrufus.springdsl.runtimeconfig
+package com.github.wakingrufus.funk
 
 import com.github.wakingrufus.funk.base.getDsl
 import io.github.oshai.kotlinlogging.KotlinLogging
