+++ spring-funk-runtimeconfig/src/test/kotlin/com/github/wakingrufus/funk/IntegrationTest.kt
@@ -1,4 +1,4 @@
-package com.github.wakingrufus.springdsl.runtimeconfig
+package com.github.wakingrufus.funk
 
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.assertj.core.api.Assertions.assertThat
