+++ spring-funk-runtimeconfig/src/test/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.context.ApplicationContextInitializer=\
-com.github.wakingrufus.springdsl.runtimeconfig.TestApplication
+com.github.wakingrufus.funk.TestApplication
